+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.8.1] - 2025-02-19
+### Fixed
+- [#535](https://github.com/unity-sds/unity-data-services/pull/535) fix: sorting key order was not kept previously
+
+## [9.8.0] - 2025-02-10
+### Changed
+- [#532](https://github.com/unity-sds/unity-data-services/pull/532) feat: granules sort by time or other properties
+
+## [9.7.0] - 2025-02-10
+### Changed
+- [#521](https://github.com/unity-sds/unity-data-services/pull/521) feat: gemx branch
+
 ## [9.6.0] - 2025-02-03
 ### Changed
 - [#517](https://github.com/unity-sds/unity-data-services/pull/517) feat: stac browser oidc cookie